auto_cut/pyproject.toml
[project]
name = "auto-cut"
version = "0.1.0"
description = "智能切图工具，产生不同平台可用的壁纸。"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.54.0",
    "pillow>=12.0.0",
]
